requirements.txt
# ── Azure SDKs ────────────────────────────────────────────────────────────────
openai>=1.30.0,<2.0.0
azure-search-documents>=11.4.0,<12.0.0
azure-core>=1.30.0,<2.0.0

# ── Backend ───────────────────────────────────────────────────────────────────
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
pydantic>=2.7.0,<3.0.0

# ── Frontend ──────────────────────────────────────────────────────────────────
streamlit>=1.35.0,<2.0.0
requests>=2.31.0,<3.0.0

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0       # async test client for FastAPI
